// src/app/modules/hcm/shared/models/masters/job-position.ts
export class JobPosition {
    id: number = -1;
    company: number = -1;
    hierarchicalLevel: number = -1;
    mainJobPosition: number = -1;
    mTJobPosition: number = -1;
    fixedPositions: number = -1;
    temporaryPositions: number = -1;
    name: string = "";
    description: string = "";
	duties: string = "";
	goals: string = "";
	riskAnalysis: string = "";
    supervisoryPosition: number = -1;
}